0.1.0

+++ core/types/model.go
@@ -1,10 +1,10 @@
 package types
 
 type Pagination struct {
-	Total      int64 `json:"total"`
-	Page       int   `json:"page"`
-	PageSize   int   `json:"page_size"`
-	TotalPages int   `json:"total_pages"`
+	Total      int `json:"total"`
+	Page       int `json:"page"`
+	PageSize   int `json:"page_size"`
+	TotalPages int `json:"total_pages"`
 }
 
 type PaginatedResponse struct {
